refactor post component style and unique key error
