+++ digital-task/src/components/project-posts/PostWrapper.tsx
@@ -6,6 +6,8 @@ import Search from "../search/Search";
 export default function PostWrapper() {
   const { filteredData, setSearchTerm } = useFetchPost();
 
+  console.log(filteredData, "filter");
+
   const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     setSearchTerm(event.target.value);
   };
@@ -24,8 +26,8 @@ export default function PostWrapper() {
     <div className="post-container">
       <Search onSearch={handleSearchChange} />
       <div className="post-card-wrapper">
-        {filteredData?.map((item) => (
-          <PostCard key={item.id} post={item} onClick={openModal} />
+        {filteredData?.map((item, index) => (
+          <PostCard key={index} post={item} onClick={() => openModal(item)} />
         ))}
       </div>
       {selectedPost && (
